+++ src/main/java/com/emc/mongoose/core/impl/load/model/BasicItemProducer.java
@@ -20,10 +20,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.locks.Condition;
-import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.LockSupport;
-import java.util.concurrent.locks.ReentrantLock;
 
 /**
  Created by kurila on 19.06.15.
@@ -43,21 +40,21 @@ implements ItemProducer<T> {
 	protected T lastDataItem;
 	protected boolean isCircular;
 	protected boolean isShuffling;
-	protected long maxItemQueueSize;
+	protected int maxItemQueueSize;
 	//
 	protected volatile boolean areAllItemsProduced = false;
 	protected volatile long producedItemsCount = 0;
 	//
 	protected BasicItemProducer(
 		final ItemSrc<T> itemSrc, final long maxCount, final int batchSize,
-		final boolean isCircular, final boolean isShuffling, final long maxItemQueueSize
+		final boolean isCircular, final boolean isShuffling, final int maxItemQueueSize
 	) {
 		this(itemSrc, maxCount, batchSize, isCircular, isShuffling, maxItemQueueSize, 0, null);
 	}
 	//
 	private BasicItemProducer(
 		final ItemSrc<T> itemSrc, final long maxCount, final int batchSize,
-		final boolean isCircular, final boolean isShuffling, final long maxItemQueueSize,
+		final boolean isCircular, final boolean isShuffling, final int maxItemQueueSize,
 		final long skipCount, final T lastDataItem
 	) {
 		this.itemSrc = itemSrc;
@@ -68,8 +65,7 @@ implements ItemProducer<T> {
 		this.isCircular = isCircular;
 		this.isShuffling = isShuffling;
 		this.maxItemQueueSize = maxItemQueueSize;
-		//
-		this.uniqueItems = new ConcurrentHashMap<>((int) maxItemQueueSize);
+		this.uniqueItems = new ConcurrentHashMap<>(maxItemQueueSize);
 	}
 	//
 	@Override
